Moved step functions into stepStore (as actions)
