+++ src/components/CoBDComponents/CodbFooter.vue
@@ -1,33 +1,11 @@
 <script setup lang="ts">
 import { useStepStore } from '../../stores/stepStore'
-import { useReportStore } from '@/stores/reportStore'
 import { storeToRefs } from 'pinia'
 import FormButton from '../FormComponents/FormButton.vue'
-import { useRouter } from 'vue-router'
 
-const reportStore = useReportStore()
-const { inputValid } = storeToRefs(reportStore)
-const { setBlankSubmitErrorAction } = reportStore
 const stepStore = useStepStore()
 const { stepCurrent } = storeToRefs(stepStore)
-
-const router = useRouter()
-
-const backStepFunction = () => {
-  const backStep = router.currentRoute.value.meta.previous
-  router.push(backStep)
-}
-
-const nextStepFunction = () => {
-  const nextStep = router.currentRoute.value.meta.next
-  
-  if (!inputValid.value) {
-    setBlankSubmitErrorAction('Please entered all required information')
-  } else if (inputValid.value) {
-    setBlankSubmitErrorAction('')
-    router.push(nextStep)
-  }
-}
+const { nextStepAction, backStepAction } = stepStore
 
 </script>
 <template>
@@ -38,13 +16,13 @@ const nextStepFunction = () => {
     <div>
       <form-button
         label="Back"
-        @click="backStepFunction()"
+        @click="backStepAction()"
         class="btn-back"
         v-show="stepCurrent !== '/' || '/company-name-step'"
       />
       <form-button
         :label="stepCurrent === '/' ? 'Let\'s Go!' : 'Next'"
-        @click="nextStepFunction()"
+        @click="nextStepAction()"
         class="btn-next mb-20 sm:mb-0 bg-nextButtonArrow bg-no-repeat bg-buttonArrow bg-15%"
         type="submit"
       />
