+++ src/stores/reportStore.ts
@@ -7,33 +7,34 @@ import formatMoney from '../assets/utility_functions/formatMoney'
 
 export const useReportStore = defineStore('reportStore', {
   state: (): reportState => ({
-    // companyName: 'PhotoBomb',
-    // bookingsPerMonth: 3,
-    // priceAveragePerBooking: 1000.00,
-    // hoursAveragePerBooking: 200,
-    // variableCosts: [{ id: 2, name: 'Parking', category: 'Overhead', amount: 80 }] as VariableCostObj[],
-    // totalVariableCosts: 80.00,
-    // fixedCosts: [
-    //   { id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  }
-    // ] as FixedCostObj[],
-    // totalFixedCosts: 1200.00,
-    // payPerMonth: 200.00,
-    // savingsPerMonth: 200.00,
-    // userEmail: 'e@e.com',
-    // inputValid: true,
-    companyName: '',
-    bookingsPerMonth: 0,
-    priceAveragePerBooking: 0.00,
-    hoursAveragePerBooking: 0,
-    variableCosts: [] as VariableCostObj[],
-    totalVariableCosts: 0.00,
-    fixedCosts: [ { id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  }, { id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  }, ] as FixedCostObj[],
-    totalFixedCosts: 0.00,
-    payPerMonth: 0.00,
-    savingsPerMonth: 0.00,
-    userEmail: '',
+    companyName: 'PhotoBomb',
+    bookingsPerMonth: 3,
+    priceAveragePerBooking: 1000.00,
+    hoursAveragePerBooking: 200,
+    variableCosts: [{ id: 2, name: 'Parking', category: 'Overhead', amount: 80 }] as VariableCostObj[],
+    totalVariableCosts: 80.00,
+    fixedCosts: [
+      { id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  }, { id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 1, name: 'Rent', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 2, name: 'Test2', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  },{ id: 3, name: 'Test3', category: 'Overhead', amount: 1000, payPeriod: 'monthly', individualTotal: 1000  }, 
+    ] as FixedCostObj[],
+    totalFixedCosts: 1200.00,
+    payPerMonth: 200.00,
+    savingsPerMonth: 200.00,
+    userEmail: 'e@e.com',
     inputValid: true,
     blankSubmitError: '',
+    // companyName: '',
+    // bookingsPerMonth: 0,
+    // priceAveragePerBooking: 0.00,
+    // hoursAveragePerBooking: 0,
+    // variableCosts: [] as VariableCostObj[],
+    // totalVariableCosts: 0.00,
+    // fixedCosts: [] as FixedCostObj[],
+    // totalFixedCosts: 0.00,
+    // payPerMonth: 0.00,
+    // savingsPerMonth: 0.00,
+    // userEmail: '',
+    // inputValid: true,
+    // blankSubmitError: '',
   }),
   actions: {
     addCompanyNameAction(companyName: string) {
